Work with CTEST 3

+++ src/deposit.h
@@ -1,7 +1,7 @@
 #ifndef H_DEPOSIT
 #define H_DEPOSIT
 
-float prover(float vklad, int day);
-int corect(float vklad, int day);
+int prover(int vklad, int day);
+int corect(int vklad, int day);
 
 #endif
